chore: use message class in specs and categorize request specs
